functions/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
assemblyai==0.34.0
azure-core==1.31.0
azure-functions==1.21.1
azure-identity==1.18.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.8.0
azure-keyvault-keys==4.9.0
azure-keyvault-secrets==4.8.0
bump2version==1.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isodate==0.6.1
msal==1.31.0
msal-extensions==1.2.0
portalocker==2.10.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
